requirements.txt
flask 
requests 
python-dotenv 
openai 
pillow
beautifulsoup4
google-generativeai
google-genai